Moved all source files into src
