+++ App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:midi_sequencer"
-             StartupUri="/windows/GeneralWindow.xaml">
+             StartupUri="/src/windows/GeneralWindow.xaml">
     <Application.Resources>
          
     </Application.Resources>
