+++ src/view/component/general/ChannelSettings.cs
@@ -1,0 +1,18 @@
+﻿using midi_sequencer.src.view.component;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Controls;
+
+namespace midi_sequencer.src.view.component.general
+{
+    internal class ChannelSettings : Component
+    {
+        public Grid Build()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
