+++ src/windows/GeneralWindow.xaml
@@ -1,0 +1,12 @@
+﻿<Window x:Class="midi_sequencer.windows.GeneralWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:midi_sequencer.windows"
+        mc:Ignorable="d"
+        Title="MIDI sequencer" Height="450" Width="800" MinWidth="800" MinHeight="459">
+    <Grid>
+        
+    </Grid>
+</Window>
