+++ src/windows/GeneralWindow.xaml.cs
@@ -1,0 +1,32 @@
+﻿using midi_sequencer.src.view;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+
+namespace midi_sequencer.windows
+{
+    /// <summary>
+    /// Логика взаимодействия для GeneralWindow.xaml
+    /// </summary>
+    public partial class GeneralWindow : Window
+    {
+        private DesignManager designManager;
+
+        public GeneralWindow()
+        {
+            InitializeComponent();
+            this.designManager = new(this);
+            designManager.GeneralWindow();
+        }
+    }
+}
