+++ src/windows/PianoRollWindow.xaml
