+++ src/windows/PianoRollWindow.xaml.cs
