+++ src/windows/PlaybackWindow.xaml
