+++ src/windows/PlaybackWindow.xaml.cs
