+++ src/windows/icon.png
