--- view/component/Component.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Controls;
-
-namespace midi_sequencer.view.component
-{
-    internal interface Component
-    {
-        public Grid Build();
-    }
-}
