--- view/component/general/ChannelSettings.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Controls;
-
-namespace midi_sequencer.view.component.general
-{
-    internal class ChannelSettings : Component
-    {
-        public Grid Build()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
